pyproject.toml
[project]
name = "financeportfolio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.4.3",
    "fredapi>=0.5.2",
    "hmmlearn>=0.3.3",
    "langchain>=1.1.3",
    "langchain-anthropic>=1.2.0",
    "langgraph>=1.0.4",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "scipy>=1.16.3",
    "streamlit>=1.52.1",
    "tenacity>=9.1.2",
    "yfinance>=0.2.66",
]

[dependency-groups]
dev = [
    "bandit>=1.9.2",
    "isort>=7.0.0",
    "pyrefly>=0.45.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.8",
    "xenon>=0.9.3",
]

[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true
force_single_line = false
combine_as_imports = true
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
known_first_party = ["financeportfolio"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "W",      # pycodestyle warnings
    "S",      # flake8-bandit (security)
    "C901",   # mccabe complexity
    "B",      # flake8-bugbear
]
# Disable isort rules in ruff - isort is the authority
ignore = ["I"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]  # Allow assert in tests

[tool.bandit]
exclude_dirs = ["tests", ".venv", "venv"]
skips = ["B101"]  # Skip assert warnings (tests use asserts)
